+++ lib/config/config.dart
@@ -1,3 +1,4 @@
 export 'canceller.dart';
+export 'impak_retro_client_interceptor.dart';
 export 'request_methods.dart';
 export 'time_units.dart';