+++ lib/impak.dart
@@ -1,3 +1,4 @@
+export 'export/impak_http.dart';
 export 'impak_retro.dart';
 export 'models/models.dart';
 export 'config/config.dart';